refactor(ir): separate resolved middle-end preparation from source services
Extract the existing GVN and async transformations behind explicit controls
and transaction-owned counters. Preserve legacy environment conversion,
import-time diagnostics, pass order, callable identity and failure accounting.
Route derived identity construction through the canonical value contract.

Add 77 focused controls and record the four existing failures reproduced on
the unchanged prerequisite. Detached source capture and typed preparation
remain the next checkpoint; this does not certify retirement or lift holds.

Model: Codex GPT-6 Astra High
Implementation-model: Codex GPT-6 Astra Low
Architecture-model: Codex GPT-6 Astra High
